BUG: Fix ResampleAnImage example (Python3)
Since Python 3, the division operation returns a float,
even if both operands are integers.

The setitem method from the Index class expects a long type,
so explicitely casting the result of the division to int allows
to run the example with Python3.

+++ src/Filtering/ImageGrid/ResampleAnImage/Code.py
@@ -46,8 +46,8 @@ size = inputImage.GetLargestPossibleRegion().GetSize()
 spacing = inputImage.GetSpacing()
 
 centralPixel = itk.Index[Dimension]()
-centralPixel[0] = size[0] / 2
-centralPixel[1] = size[1] / 2
+centralPixel[0] = int(size[0] / 2)
+centralPixel[1] = int(size[1] / 2)
 centralPoint = itk.Point[ScalarType, Dimension]()
 centralPoint[0] = centralPixel[0]
 centralPoint[1] = centralPixel[1]
